Renames index to solr_index for clarity's sake
