+++ core/management/commands/index.py
@@ -3,7 +3,7 @@ import logging
 from django.core.management.base import BaseCommand
     
 from core.management.commands import configure_logging
-from core.index import index_titles, index_pages
+from core.solr_index import index_titles, index_pages
 
 configure_logging("index_logging.config", "index.log")
