+++ core/management/commands/index_pages.py
@@ -1,7 +1,7 @@
 from django.core.management.base import BaseCommand
     
 from core.management.commands import configure_logging
-from core.index import index_pages
+from core.solr_index import index_pages
 
 configure_logging("index_pages_logging.config", "index_pages.log")
