+++ core/solr_index.py
